debian-bullseye/nextpnr: disable parallel make
GitHub Actions is annoyingly failing to build nextpnr/ice40 on Debian
Bullseye with the following message:

[debian/bullseye] nextpnr · ice40 icestorm
Received request to deprovision: The request was cancelled by the remote
provider.

No further logs are produced.

However, it does work on Debian Buster and nextpnr/ecp5 does work
on both Buster and Bullseye.
Vague references in the internet mentioned it might be related to the
host (Azure) and to resource allocation.
Suprisingly, executing make without '-j' works.